refactor: 상품을 등록하는 로직 수정 및 DIP 적용
- 비속어를 검증하는 메서드 생성
- 비속어를 응용 서비스 모듈에 위치시켜서 DIP 적용
